+++ packages/authentication/modules/profile/index.ts
@@ -1,0 +1,5 @@
+export { default as useCurrentProfile } from './useCurrentProfile'
+export type { MinimalProfile } from '../../types/profile'
+
+export { InitialProfileProvider } from './useCurrentProfile/__tests__/__utils__/InitialProfileProvider'
+export type { InitialProfileProp } from './useCurrentProfile/__tests__/__utils__/InitialProfileProvider'