+++ packages/components/__generated__/UserProfileQuery.graphql.ts
@@ -1,5 +1,5 @@
 /**
- * @generated SignedSource<<f3527d4c54b58cbec40354c7eedb4f9f>>
+ * @generated SignedSource<<6a78928961965b4807c4cd830f7dd703>>
  * @lightSyntaxTransform
  * @nogrep
  */
@@ -17,7 +17,7 @@ export type UserProfileQuery$data = {
     | {
         readonly profile:
           | {
-              readonly ' $fragmentSpreads': FragmentRefs<'ProfileItemInlineFragment'>
+              readonly ' $fragmentSpreads': FragmentRefs<'ProfileItemFragment'>
             }
           | null
           | undefined
@@ -32,35 +32,12 @@ export type UserProfileQuery = {
 
 const node: ConcreteRequest = (function () {
   var v0 = {
-      alias: null,
-      args: null,
-      kind: 'ScalarField',
-      name: 'id',
-      storageKey: null,
-    },
-    v1 = {
-      alias: null,
-      args: null,
-      kind: 'ScalarField',
-      name: 'name',
-      storageKey: null,
-    },
-    v2 = [
-      {
-        alias: null,
-        args: null,
-        kind: 'ScalarField',
-        name: 'url',
-        storageKey: null,
-      },
-    ],
-    v3 = {
-      alias: null,
-      args: null,
-      kind: 'ScalarField',
-      name: 'path',
-      storageKey: null,
-    }
+    alias: null,
+    args: null,
+    kind: 'ScalarField',
+    name: 'id',
+    storageKey: null,
+  }
   return {
     fragment: {
       argumentDefinitions: [],
@@ -85,51 +62,9 @@ const node: ConcreteRequest = (function () {
               plural: false,
               selections: [
                 {
-                  kind: 'InlineDataFragmentSpread',
-                  name: 'ProfileItemInlineFragment',
-                  selections: [
-                    v0 /*: any*/,
-                    v1 /*: any*/,
-                    {
-                      alias: null,
-                      args: [
-                        {
-                          kind: 'Variable',
-                          name: 'height',
-                          variableName: 'avatarSize',
-                        },
-                        {
-                          kind: 'Variable',
-                          name: 'width',
-                          variableName: 'avatarSize',
-                        },
-                      ],
-                      concreteType: 'File',
-                      kind: 'LinkedField',
-                      name: 'image',
-                      plural: false,
-                      selections: v2 /*: any*/,
-                      storageKey: null,
-                    },
-                    {
-                      alias: null,
-                      args: null,
-                      concreteType: 'URLPath',
-                      kind: 'LinkedField',
-                      name: 'urlPath',
-                      plural: false,
-                      selections: [v3 /*: any*/],
-                      storageKey: null,
-                    },
-                  ],
                   args: null,
-                  argumentDefinitions: [
-                    {
-                      defaultValue: 100,
-                      kind: 'LocalArgument',
-                      name: 'avatarSize',
-                    },
-                  ],
+                  kind: 'FragmentSpread',
+                  name: 'ProfileItemFragment',
                 },
               ],
               storageKey: null,
@@ -164,7 +99,13 @@ const node: ConcreteRequest = (function () {
               plural: false,
               selections: [
                 v0 /*: any*/,
-                v1 /*: any*/,
+                {
+                  alias: null,
+                  args: null,
+                  kind: 'ScalarField',
+                  name: 'name',
+                  storageKey: null,
+                },
                 {
                   alias: null,
                   args: [
@@ -183,7 +124,15 @@ const node: ConcreteRequest = (function () {
                   kind: 'LinkedField',
                   name: 'image',
                   plural: false,
-                  selections: v2 /*: any*/,
+                  selections: [
+                    {
+                      alias: null,
+                      args: null,
+                      kind: 'ScalarField',
+                      name: 'url',
+                      storageKey: null,
+                    },
+                  ],
                   storageKey: 'image(height:100,width:100)',
                 },
                 {
@@ -193,7 +142,16 @@ const node: ConcreteRequest = (function () {
                   kind: 'LinkedField',
                   name: 'urlPath',
                   plural: false,
-                  selections: [v3 /*: any*/, v0 /*: any*/],
+                  selections: [
+                    {
+                      alias: null,
+                      args: null,
+                      kind: 'ScalarField',
+                      name: 'path',
+                      storageKey: null,
+                    },
+                    v0 /*: any*/,
+                  ],
                   storageKey: null,
                 },
               ],
@@ -206,16 +164,16 @@ const node: ConcreteRequest = (function () {
       ],
     },
     params: {
-      cacheID: 'a85f1b3f866dce00c22e0195d8f91a9a',
+      cacheID: '4a09b56bb39794e18b189981877fa031',
       id: null,
       metadata: {},
       name: 'UserProfileQuery',
       operationKind: 'query',
-      text: 'query UserProfileQuery {\n  me {\n    profile {\n      ...ProfileItemInlineFragment\n      id\n    }\n    id\n  }\n}\n\nfragment ProfileItemInlineFragment on Profile {\n  id\n  name\n  image(width: 100, height: 100) {\n    url\n  }\n  urlPath {\n    path\n    id\n  }\n}\n',
+      text: 'query UserProfileQuery {\n  me {\n    profile {\n      ...ProfileItemFragment\n      id\n    }\n    id\n  }\n}\n\nfragment ProfileItemFragment on Profile {\n  id\n  name\n  image(width: 100, height: 100) {\n    url\n  }\n  urlPath {\n    path\n    id\n  }\n}\n',
     },
   }
 })()
 
-;(node as any).hash = 'fe5b9485e3ef19ae48b2142a3c72c7f7'
+;(node as any).hash = '8febc9fbb812c3456d12d0ff6a69c90a'
 
 export default node